tarodb: add implementation of CommitmentSelector for assets store
In this commit, we add an implementation of CommitmentSelector for the
main assets store. Along the way, we extend the new QueryAsset method to
allow us to select all the assets anchored to a given outpoint. We can
then use this to fetch all the other assets that are committed alongside
a main asset, as we need to also return the full taro commitment for
each selected asset.

This implementation will be used to do proper coin selection for the
upcoming batched send logic.