tarofreighter: add new ExportLog interface for committing spends
In this commit, we add a new `ExportLog` interface that'll be used to
update the on-disk information related to which assets are spent and how
they're spent.

When an asset is spent, all assets that were collocated in the same
commitment must now be re-anchored in a new commitment. However _some_
assets will also be mutated if they were used to create a split, or mark
a complete spend/send to a party (interactive or non-interactive).

A spend moves from pending to confirmed once it's sufficiently buried in
the chain. At this point, the on-disk information that tracks where the
chain_txn is located will be updated.

At this time, roll backs aren't supported (re-org or need to undo a
spend). As a result, a spend should only be marked as pending once it's
certain that the transactions that anchors the new assets will
eventually be broadcaster and confirmed.