multi: use SerializedKey for PrevID's ScriptKey
Because the PrevID is used as a map key, we cannot have struct members
that aren't always comparable. For example the x and y coordinates of a
public key can be represented in memory in different ways, even if the
actual serialized value might be the same (see FieldVal.Normalize()).
To make sure we don't introduce very hard to find bugs, we use the
serialized format of public keys in structs that we use as map keys.